Move ensure_additional_properties_false to base modules
