+++ python/mirascope/llm/providers/openai/responses/_utils/encode.py
@@ -39,6 +39,7 @@ from .....formatting import (
 from .....messages import AssistantMessage, Message, UserMessage
 from .....tools import FORMAT_TOOL_NAME, AnyToolSchema, BaseToolkit
 from ....base import Params, _utils as _base_utils
+from ....base._utils import ensure_additional_properties_false
 from ...model_id import OpenAIModelId, model_name
 from ...model_info import NON_REASONING_MODELS
 from ...shared import _utils as _shared_utils
@@ -197,7 +198,7 @@ def _convert_tool_to_function_tool_param(tool: AnyToolSchema) -> FunctionToolPar
     """Convert a Mirascope ToolSchema to OpenAI Responses FunctionToolParam."""
     schema_dict = tool.parameters.model_dump(by_alias=True, exclude_none=True)
     schema_dict["type"] = "object"
-    _shared_utils.ensure_additional_properties_false(schema_dict)
+    ensure_additional_properties_false(schema_dict)
 
     return FunctionToolParam(
         type="function",
@@ -220,7 +221,7 @@ def _create_strict_response_format(
         ResponseFormatTextJSONSchemaConfigParam for strict structured outputs
     """
     schema = format.schema.copy()
-    _shared_utils.ensure_additional_properties_false(schema)
+    ensure_additional_properties_false(schema)
 
     response_format: ResponseFormatTextJSONSchemaConfigParam = {
         "type": "json_schema",
